Refactor map ingestion
