+++ chips/globals.py
@@ -6,7 +6,6 @@ MOUSE = 'MOUSE'
 KEYBOARD = 'KEYBOARD'
 CLOCK = 'CLOCK'
 TILE = 'TILE'
-MAP_DATA = 'MAP_DATA'
 MAP = 'MAP'
 PLAYER = 'PLAYER'
 CAMERA = 'CAMERA'
